mlops-task/requirements.txt
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.2